Refactor into constants file.
