+++ kafka-hello-world/src/main/scala/com/labs1904/hwe/consumers/ConnectionTest.scala
@@ -1,33 +1,23 @@
 package com.labs1904.hwe.consumers
 
+import com.labs1904.hwe.util.Constants._
 import com.labs1904.hwe.util.Util
 import net.liftweb.json.DefaultFormats
-import org.apache.kafka.clients.consumer.{ConsumerConfig, ConsumerRecord, ConsumerRecords, KafkaConsumer}
-import org.apache.kafka.common.serialization.StringDeserializer
+import org.apache.kafka.clients.consumer.{ConsumerRecord, ConsumerRecords, KafkaConsumer}
 
 import java.time.Duration
-import java.util.{Arrays, Properties, UUID}
+import java.util.Arrays
 
 object ConnectionTest {
-  // TODO: this is configured to use kafka running locally, change it to your cluster
-  val BootstrapServer : String = "CHANGEME"
-  val Topic: String = "hwe-kafka-connection-test"
-  val username: String = "CHANGEME"
-  val password: String = "CHANGEME"
-  //Use this for Windows
-  val trustStore: String = "src\\main\\resources\\kafka.client.truststore.jks"
-  //Use this for Mac
-  //val trustStore: String = "src/main/resources/kafka.client.truststore.jks"
-
   implicit val formats: DefaultFormats.type = DefaultFormats
 
   def main(args: Array[String]): Unit = {
     // Create the KafkaConsumer
-    val properties = getProperties(BootstrapServer)
+    val properties = Util.getConsumerProperties(BOOTSTRAP_SERVER)
     val consumer: KafkaConsumer[String, String] = new KafkaConsumer[String, String](properties)
 
     // Subscribe to the topic
-    consumer.subscribe(Arrays.asList(Topic))
+    consumer.subscribe(Arrays.asList(DEFAULT_TOPIC))
 
     while ( {
       true
@@ -43,21 +33,4 @@ object ConnectionTest {
       })
     }
   }
-
-  def getProperties(bootstrapServer: String): Properties = {
-    // Set Properties to be used for Kafka Consumer
-    val properties = new Properties
-    properties.setProperty(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServer)
-    properties.setProperty(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, classOf[StringDeserializer].getName)
-    properties.setProperty(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, classOf[StringDeserializer].getName)
-    properties.setProperty(ConsumerConfig.GROUP_ID_CONFIG, UUID.randomUUID().toString)
-
-    properties.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
-    properties.put("security.protocol", "SASL_SSL")
-    properties.put("sasl.mechanism", "SCRAM-SHA-512")
-    properties.put("ssl.truststore.location", trustStore)
-    properties.put("sasl.jaas.config", Util.getScramAuthString(username, password))
-    properties
-  }
-
 }
