+++ kafka-hello-world/src/main/scala/com/labs1904/hwe/producers/ProducerWithFaker.scala
@@ -1,31 +1,20 @@
 package com.labs1904.hwe.producers
 
-import com.labs1904.hwe.util.Util.getScramAuthString
-import org.apache.kafka.clients.producer.{Callback, KafkaProducer, ProducerConfig, ProducerRecord, RecordMetadata}
-import org.apache.kafka.common.serialization.StringSerializer
+import com.labs1904.hwe.util.Constants._
+import com.labs1904.hwe.util.Util
 import faker._
 import net.liftweb.json.DefaultFormats
-import net.liftweb.json.Serialization.write
-
-import java.util.Properties
+import org.apache.kafka.clients.producer.{Callback, KafkaProducer, ProducerRecord, RecordMetadata}
 
 case class User(name: String, username: String, email: String)
 
 object ProducerWithFaker {
   implicit val formats: DefaultFormats.type = DefaultFormats
-  val BootstrapServer : String = "CHANGEME"
-  val Topic: String = "CHANGEME"
-  val username: String = "CHANGEME"
-  val password: String = "CHANGEME"
-  //Use this for Windows
-  val trustStore: String = "src\\main\\resources\\kafka.client.truststore.jks"
-  //Use this for Mac
-  //val trustStore: String = "src/main/resources/kafka.client.truststore.jks"
 
   def main(args: Array[String]): Unit = {
 
     // Create the Kafka Producer
-    val properties = getProperties(BootstrapServer)
+    val properties = Util.getProperties(BOOTSTRAP_SERVER)
     val producer = new KafkaProducer[String, String](properties)
 
     // create n fake records to send to topic
@@ -44,7 +33,7 @@ object ProducerWithFaker {
       //val jsonString = write(user)
       val csvString = key + "," + name.replace(",","") + "," + user.email.replace(",","")
 
-      new ProducerRecord[String, String](Topic, key, csvString)
+      new ProducerRecord[String, String](DEFAULT_TOPIC, key, csvString)
     }).foreach(record => {
 
       // send records to topic
@@ -67,19 +56,5 @@ object ProducerWithFaker {
     producer.close()
   }
 
-  def getProperties(bootstrapServer: String): Properties = {
-    // Set Properties to be used for Kafka Producer
-    val properties = new Properties
-    properties.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServer)
-    properties.setProperty(ProducerConfig.ACKS_CONFIG, "1")
-    properties.setProperty(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, classOf[StringSerializer].getName)
-    properties.setProperty(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, classOf[StringSerializer].getName)
 
-    properties.put("security.protocol", "SASL_SSL")
-    properties.put("sasl.mechanism", "SCRAM-SHA-512")
-    properties.put("ssl.truststore.location", trustStore)
-    properties.put("sasl.jaas.config", getScramAuthString(username, password))
-
-    properties
-  }
 }
