+++ kafka-hello-world/src/main/scala/com/labs1904/hwe/util/Util.scala
@@ -1,6 +1,46 @@
 package com.labs1904.hwe.util
 
+import com.labs1904.hwe.util.Constants.{PASSWORD, TRUST_STORE, USERNAME}
+import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.clients.producer.ProducerConfig
+import org.apache.kafka.common.serialization.{StringDeserializer, StringSerializer}
+
+import java.util.{Properties, UUID}
+
 object Util {
+  def getProperties(bootstrapServer: String): Properties = {
+    // Set Properties to be used for Kafka Producer
+    val properties = new Properties
+    properties.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServer)
+    properties.setProperty(ProducerConfig.ACKS_CONFIG, "1")
+    properties.setProperty(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, classOf[StringSerializer].getName)
+    properties.setProperty(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, classOf[StringSerializer].getName)
+
+    properties.put("security.protocol", "SASL_SSL")
+    properties.put("sasl.mechanism", "SCRAM-SHA-512")
+    properties.put("ssl.truststore.location", Constants.TRUST_STORE)
+    properties.put("sasl.jaas.config", getScramAuthString(Constants.USERNAME, Constants.PASSWORD))
+
+    properties
+  }
+
+  def getConsumerProperties(bootstrapServer: String): Properties = {
+    // Set Properties to be used for Kafka Consumer
+    val properties = new Properties
+    properties.setProperty(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServer)
+    properties.setProperty(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, classOf[StringDeserializer].getName)
+    properties.setProperty(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, classOf[StringDeserializer].getName)
+    properties.setProperty(ConsumerConfig.GROUP_ID_CONFIG, UUID.randomUUID().toString)
+    properties.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
+
+    properties.put("security.protocol", "SASL_SSL")
+    properties.put("sasl.mechanism", "SCRAM-SHA-512")
+    properties.put("ssl.truststore.location", TRUST_STORE)
+    properties.put("sasl.jaas.config", getScramAuthString(USERNAME, PASSWORD))
+
+    properties
+  }
+
   def getScramAuthString(username: String, password: String) = {
    s"""org.apache.kafka.common.security.scram.ScramLoginModule required
    username=\"$username\"
